+++ conkit/plot/tests/test_tools.py
@@ -10,37 +10,58 @@ import unittest
 
 
 class Test(unittest.TestCase):
+
     def test_points_on_circle_1(self):
+        coords = tools.points_on_circle(0)
+        self.assertEqual([[]], coords)
+
+    def test_points_on_circle_2(self):
+        coords = tools.points_on_circle(1)
+        self.assertEqual([[1., 0.]], coords)
+
+    def test_points_on_circle_3(self):
         coords = tools.points_on_circle(2)
         self.assertEqual([2., 0.], coords[0])
         self.assertEqual([-2., 0], coords[1])
 
-    def test_points_on_circle_2(self):
+    def test_points_on_circle_4(self):
         coords = tools.points_on_circle(3)
         self.assertEqual([3., 0.], coords[0])
         self.assertEqual([-1.5, 2.598076], coords[1])
         self.assertEqual([-1.5, -2.598076], coords[2])
 
-    def test_points_on_circle_3(self):
+    def test_points_on_circle_5(self):
         coords = tools.points_on_circle(4)
         self.assertEqual([4., 0.], coords[0])
         self.assertEqual([0., 4.], coords[1])
         self.assertEqual([-4., 0.], coords[2])
         self.assertEqual([0, -4], coords[3])
 
-    def test_points_on_circle_4(self):
+    def test_points_on_circle_6(self):
         coords = tools.points_on_circle(4, h=2)
         self.assertEqual([6., 0.], coords[0])
         self.assertEqual([2., 4.], coords[1])
         self.assertEqual([-2., 0.], coords[2])
         self.assertEqual([2, -4], coords[3])
 
-    def test_points_on_circle_5(self):
+    def test_points_on_circle_7(self):
         coords = tools.points_on_circle(4, k=-3)
         self.assertEqual([4., -3.], coords[0])
         self.assertEqual([0., 1.], coords[1])
         self.assertEqual([-4., -3.], coords[2])
         self.assertEqual([0, -7], coords[3])
+
+    def test_radius_around_circle_1(self):
+        p1, p2 = [0., 0.], [0., 0]
+        self.assertEqual(0.0, tools.radius_around_circle(p1, p2)) 
+
+    def test_radius_around_circle_2(self):
+        p1, p2 = [2., 0.], [-2., 0]
+        self.assertEqual(1.6, tools.radius_around_circle(p1, p2)) 
+
+    def test_radius_around_circle_3(self):
+        p1, p2 = [2., 1.], [-2., 4]
+        self.assertAlmostEqual(2.0, tools.radius_around_circle(p1, p2)) 
 
     def test__isinstance_1(self):
         h = Contact(1, 2, 2)